refactor(backend): extract card managed media owner
